raise MosaicExists error when not forcing overwrite

+++ cogeo_mosaic/backends/file.py
@@ -1,6 +1,7 @@
 """cogeo-mosaic File backend."""
 
 import json
+import os
 from typing import List
 
 import attr
@@ -14,7 +15,7 @@ from cogeo_mosaic.backends.utils import (
     get_assets_from_json,
 )
 from cogeo_mosaic.cache import lru_cache
-from cogeo_mosaic.errors import _FILE_EXCEPTIONS, MosaicError
+from cogeo_mosaic.errors import _FILE_EXCEPTIONS, MosaicError, MosaicExists
 from cogeo_mosaic.mosaic import MosaicJSON
 
 
@@ -35,8 +36,11 @@ class FileBackend(BaseBackend):
             self.mosaic_def.tiles, self.quadkey_zoom, tile.x, tile.y, tile.z
         )
 
-    def write(self, gzip: bool = None):
+    def write(self, overwrite: bool = False, gzip: bool = None):
         """Write mosaicjson document to a file."""
+        if not overwrite and os.path.exists(self.path):
+            raise MosaicExists("Mosaic file already exist, use `overwrite=True`.")
+
         body = self.mosaic_def.dict(exclude_none=True)
         with open(self.path, "wb") as f:
             try:
